Refactoring and implement thread in factory method

+++ src/main/java/it/unibo/pensilina14/bullet/ballet/model/effects/EffectsUtils.java
@@ -56,7 +56,7 @@ public class EffectsUtils {
 			this.value = value;
 		}
 		
-		public final double getValue() {
+		public final long getValue() {
 			return this.value;
 		}
 	}
@@ -85,7 +85,7 @@ public class EffectsUtils {
 			this.value = value;
 		}
 		
-		public final double getValue() {
+		public final long getValue() {
 			return this.value;
 		}
 	}
